Update code and build to allow GNUstep build
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@4270 d25b4930-3292-e311-b074-0010c6de5da8

+++ ported-apps/Util/Bean/KBWordCountingTextStorage.m
@@ -159,7 +159,7 @@ NSString *KBTextStorageStatisticsDidChangeNotification = @"KBTextStorageStatisti
 	return [text string];
 }
 
-- (NSDictionary *)attributesAtIndex:(unsigned)index effectiveRange:(NSRangePointer)aRange
+- (NSDictionary *)attributesAtIndex:(NSUInteger)index effectiveRange:(NSRangePointer)aRange
 {
 	return [text attributesAtIndex:index effectiveRange:aRange];
 }
